wifi: extend feature flag from int to long
WiFi features run out of all 32 bits of int.
Extend feature flag from int to long for more bits.

Bug: 122939035
Test: build
Test: atest frameworks/opt/net/wifi/tests/wifitests

+++ wifi/java/android/net/wifi/WifiManager.java
@@ -2055,7 +2055,7 @@ public class WifiManager {
     /** @hide */
     public static final int WIFI_FEATURE_DPP              = 0x80000000; // DPP (Easy-Connect)
 
-    private int getSupportedFeatures() {
+    private long getSupportedFeatures() {
         try {
             return mService.getSupportedFeatures();
         } catch (RemoteException e) {
@@ -2063,7 +2063,7 @@ public class WifiManager {
         }
     }
 
-    private boolean isFeatureSupported(int feature) {
+    private boolean isFeatureSupported(long feature) {
         return (getSupportedFeatures() & feature) == feature;
     }
     /**
